requirements.txt
fastapi
uvicorn
numpy
openai
rapidfuzz
serpapi
python-dotenv
beautifulsoup4
google-serp-api
google-search-results